% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kitten_growth}
\alias{kitten_growth}
\title{Jeannine's kitten growth data.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{kitten_id}{unique id of kitten}
\item{sex}{sex of kittens}
\item{total_mass}{total mass}
\item{bag}{mass of bag}
\item{mass}{mass of kitten}
\item{id_features}{kitten features}
\item{name}{kitten name}
\item{datetime}{kitten datetime}
...
}
}
\usage{
kitten_growth
}
\description{
A dataset tracking the growth of Jeannine's kittens.
}
\keyword{datasets}
