% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_kitten_growth}
\alias{plot_kitten_growth}
\title{Plot kitten growth}
\usage{
plot_kitten_growth(kittens = purrfectdata::kitten_growth)
}
\arguments{
\item{kittens}{A data.frame of the kitten growth.}
}
\value{
A ggplot2 object
}
\description{
Plot kitten growth
}
\examples{
plot_kitten_growth()
}
